chore: rename repository to accs-daemon
