+++ src/accs_agent/__init__.py
@@ -1,0 +1,1 @@
+"""Service agent package."""
