--- src/accs_app/__init__.py
@@ -1,1 +1,0 @@
-"""ACCScore application package."""
