--- src/accs_app/agents/__init__.py
@@ -1,1 +1,0 @@
-"""Agent components for ACCS application."""
