--- src/accs_app/db/__init__.py
@@ -1,1 +1,0 @@
-"""Database helpers for the builder application."""
