+++ src/accs_builder_daemon/__init__.py
@@ -1,0 +1,1 @@
+"""Builder daemon package."""
