+++ src/accs_builder_daemon/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Domain models for the builder daemon."""
