+++ src/accs_builder_daemon/repo/__init__.py
@@ -1,0 +1,1 @@
+"""Repository implementations for the builder daemon."""
