+++ src/accs_infra/__init__.py
@@ -1,0 +1,1 @@
+"""Shared infrastructure components."""
